# swagger/spring-server/src/main/java/ch/heig/amt/gamification/api/endpoints/BadgesApiController.java
package ch.heig.amt.gamification.api.endpoints;

import ch.heig.amt.gamification.api.BadgesApi;
import ch.heig.amt.gamification.api.model.Badge;
import ch.heig.amt.gamification.entities.BadgeEntity;
import ch.heig.amt.gamification.repositories.BadgeRepository;
import io.swagger.annotations.ApiParam;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.servlet.support.ServletUriComponentsBuilder;

import javax.validation.Valid;
import java.net.URI;
import java.util.ArrayList;
import java.util.List;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-07-26T19:36:34.802Z")

@Controller
public class BadgesApiController implements BadgesApi {

    @Autowired
    BadgeRepository badgeRepository;

    @Override
    public ResponseEntity<Object> createBadge(@ApiParam(value = "", required = true) @Valid @RequestBody Badge badge) {
        // Registration of the badge as an entity
        BadgeEntity newBadgeEntity = toBadgeEntity(badge);
        badgeRepository.save(newBadgeEntity);

        // Get the badge and build the response content of this badge from his new link with id
        URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
                .buildAndExpand(newBadgeEntity.getBadgeId()).toUri();
        return ResponseEntity.created(location).build();
    }

    @Override
    public ResponseEntity<Object> deleteBadge(@PathVariable Integer badgeId) {
        BadgeEntity badgeEntity = badgeRepository.findOne(badgeId);

        // Checking if existing badge
        if (badgeEntity == null) {
            return ResponseEntity.notFound().build();
        }

        // delete badge and send no content as response
        badgeRepository.delete(badgeEntity);
        return ResponseEntity.noContent().build();
    }

    @Override
    public ResponseEntity<Object> deleteBadges() {
        for (BadgeEntity badgeEntity : badgeRepository.findAll()) {
            badgeRepository.delete(badgeEntity);
        }
        return ResponseEntity.noContent().build();
    }

    @Override
    public ResponseEntity<Object> editBadge(@RequestBody Badge badge, @PathVariable Integer badgeId) {
        BadgeEntity badgeEntity = badgeRepository.findOne(badgeId);

        // Checking if existing badge
        if (badgeEntity == null) {
            return ResponseEntity.notFound().build();
        }

        // edit badge and send no content as response
        badge.setBadgeName(badge.getBadgeName());
        badgeRepository.save(toBadgeEntity(badge));
        return ResponseEntity.accepted().build();
    }

    @Override
    public ResponseEntity<Badge> getBadge(@PathVariable Integer badgeId) {
        BadgeEntity badgeEntity = badgeRepository.findOne(badgeId);

        // Checking if existing badge
        if (badgeEntity == null) {
            return ResponseEntity.notFound().build();
        }

        return ResponseEntity.ok(toBadge(badgeEntity));
    }

    @Override
    public ResponseEntity<List<Badge>> getBadges() {
        List<Badge> badges = new ArrayList<>();
        for (BadgeEntity badgeEntity : badgeRepository.findAll()) {
            badges.add(toBadge(badgeEntity));
        }
        return ResponseEntity.ok(badges);
    }

    public BadgeEntity toBadgeEntity(Badge badge) {
        BadgeEntity entity = new BadgeEntity();
        entity.setBadgeId(badge.getBadgeId());
        entity.setBadgeName(badge.getBadgeName());
        return entity;
    }

    private Badge toBadge(BadgeEntity entity) {
        Badge badge = new Badge();
        badge.setBadgeId((entity.getBadgeId()));
        badge.setBadgeName(entity.getBadgeName());
        return badge;
    }
}